Moves weapon pages to a dedicated folder
Relocates the weapon-related pages (Create, Delete, Edit, and List) to a dedicated "Weapons" folder within the Pages directory.
This change improves project organization and maintainability by grouping related components.
