+++ src/WebApiLearning.Blazor/Components/Pages/Weapons/DeleteWeapon.razor
