+++ src/WebApiLearning.Blazor/Components/Pages/Weapons/EditWeapon.razor
