+++ src/WebApiLearning.Blazor/Components/Pages/Weapons/Weapons.razor
