+++ src/WebApiLearning.Blazor/Components/Pages/Weapons/Weapons.razor.cs
@@ -1,6 +1,6 @@
 using WebApiLearning.Domain.Entities;
 
-namespace WebApiLearning.Blazor.Components.Pages;
+namespace WebApiLearning.Blazor.Components.Pages.Weapons;
 
 public partial class Weapons
 {
